Revising changes by PR reviews
